Move InfoCardDetails to InfoCard Component Also have InfoCard only have that as an input
